requirements.txt
Django==1.11
djangorestframework==3.6.3
psycopg2==2.7.1
coverage==4.4.1
pytest==3.1.2
pytest-cov==2.5.1
pytest-django==3.1.2
python-dotenv==0.6.4
